# openapi.yaml
openapi: 3.0.3
info:
  title: FreeSWITCH Call Control API
  description: |
    RESTful API for controlling FreeSWITCH calls and managing mod_callcenter
    resources (queues, agents, tiers, members) via Event Socket Library (ESL).

    ## Authentication

    When `FSAPI_AUTH_TOKENS` is configured, all requests (except from localhost)
    must include a valid Bearer token in the `Authorization` header.

    ## Multi-Tenant Authorization

    The `X-Allowed-Contexts` header enables per-tenant access control:
    - **Omitted or empty** — unrestricted access (backward compatible)
    - **`*`** — explicit unrestricted / super-admin access
    - **`domain1.com,domain2.com`** — restrict to listed domains

    For call endpoints the domain is matched against the call's `accountcode`.
    For callcenter endpoints it is matched against the domain portion of
    queue/tier names (`name@domain`) or the `domain_name=` value inside agent
    contact strings.
  version: 0.4.2
  contact:
    name: Emaktel
    url: https://github.com/emaktel/fs-api
servers:
  - url: http://localhost:37274
    description: Local development

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Required for remote requests when FSAPI_AUTH_TOKENS is configured.
        Localhost requests bypass authentication.

  # -------------------------------------------------------------------------
  # Headers
  # -------------------------------------------------------------------------
  parameters:
    XAllowedContexts:
      name: X-Allowed-Contexts
      in: header
      required: false
      description: >
        Comma-separated list of allowed FreeSWITCH contexts (domains).
        Use `*` for unrestricted access. Omit for backward-compatible
        unrestricted access.
      schema:
        type: string
        example: "customer1.example.com,customer2.example.com"
...
    QueueName:
      name: queue_name
      in: path
      required: true
      schema:
        type: string
      description: "Queue name in `name@domain` format"
      example: "support@customer1.example.com"
...
  headers:
    X-Request-ID:
      description: Unique identifier for the request, useful for log correlation
      schema:
        type: string
        format: uuid

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          additionalProperties: true
      required: [status, data]
...
  responses:
...
    BadGateway:
      description: FreeSWITCH ESL command error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    ServiceUnavailable:
      description: FreeSWITCH ESL connection unavailable
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"

# ===========================================================================
# Paths
# ===========================================================================
paths:
...
  /v1/status:
    get:
      tags: [Status]
      summary: Get FreeSWITCH status
      operationId: getStatus
      parameters:
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: FreeSWITCH status retrieved
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # -------------------------------------------------------------------------
  # Registrations
  # -------------------------------------------------------------------------
...
  /v1/callcenter/queues:
    get:
      tags: [Callcenter - Queues]
      summary: List all queues
      description: >
        Returns all callcenter queues. When access is restricted the list is
        filtered to queues whose `name` field domain matches an allowed context.
      operationId: ccListQueues
      parameters:
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queues retrieved
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCListResponse"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/callcenter/queues/count:
    get:
      tags: [Callcenter - Queues]
      summary: Count queues
      description: >
        Returns the total number of queues. For restricted access the count
        is derived from a filtered list.
      operationId: ccCountQueues
      parameters:
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue count
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCCountResponse"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/agents:
    get:
      tags: [Callcenter - Queues]
      summary: List agents in a queue
      operationId: ccListQueueAgents
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue agents retrieved
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/agents/count:
    get:
      tags: [Callcenter - Queues]
      summary: Count agents in a queue
      description: >
        Returns the number of agents in the queue. Supports an optional
        `status` query parameter to count only agents with a specific status.
      operationId: ccCountQueueAgents
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter agents by status (e.g. Available, Logged Out)
      responses:
        "200":
          description: Agent count
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCCountResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/members:
    get:
      tags: [Callcenter - Queues]
      summary: List members (callers) in a queue
      operationId: ccListQueueMembers
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue members retrieved
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/members/count:
    get:
      tags: [Callcenter - Queues]
      summary: Count members in a queue
      operationId: ccCountQueueMembers
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Member count
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCCountResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/tiers:
    get:
      tags: [Callcenter - Queues]
      summary: List tiers in a queue
      operationId: ccListQueueTiers
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue tiers retrieved
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/tiers/count:
    get:
      tags: [Callcenter - Queues]
      summary: Count tiers in a queue
      operationId: ccCountQueueTiers
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Tier count
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCCountResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/load:
    post:
      tags: [Callcenter - Queues]
      summary: Load a queue into memory
      operationId: ccLoadQueue
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue loaded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/unload:
    post:
      tags: [Callcenter - Queues]
      summary: Unload a queue from memory
      operationId: ccUnloadQueue
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue unloaded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/callcenter/queues/{queue_name}/reload:
    post:
      tags: [Callcenter - Queues]
      summary: Reload a queue configuration
      operationId: ccReloadQueue
      parameters:
        - $ref: "#/components/parameters/QueueName"
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Queue reloaded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"

  # -------------------------------------------------------------------------
  # Callcenter — Agents
  # -------------------------------------------------------------------------
  /v1/callcenter/agents:
    get:
      tags: [Callcenter - Agents]
      summary: List all agents
      description: >
        Returns all callcenter agents. When access is restricted the list is
        filtered by extracting `domain_name=` from each agent's `contact` field.
      operationId: ccListAgents
      parameters:
        - $ref: "#/components/parameters/XAllowedContexts"
      responses:
        "200":
          description: Agents retrieved
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCListResponse"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      tags: [Callcenter - Agents]
      summary: Add a new agent
      operationId: ccAddAgent
      parameters:
        - $ref: "#/components/parameters/XAllowedContexts"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentAddRequest"
      responses:
        "200":
          description: Agent added
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
...